+++ CHANGELOG.md
@@ -7,6 +7,13 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ## [Unreleased]
 
+## [0.5.0] - 2026-04-13
+
+### Added
+- `stateStream` property returning `AsyncStream<State>` for reactive state observation
+- `transitionStream` property returning `AsyncStream` of `(from, event, to)` tuples
+- Multiple concurrent subscribers supported
+
 ## [0.4.0] - 2026-04-13
 
 ### Added